split spell diff function into historical and simulated and decide at build time
