+++ modules/executive/components/SpellEffectsTab.tsx
@@ -6,8 +6,8 @@ import { Icon as DaiUIIcon } from '@makerdao/dai-ui-icons';
 import Stack from 'modules/app/components/layout/layouts/Stack';
 import SkeletonThemed from 'modules/app/components/SkeletonThemed';
 import { formatDateWithoutTime } from 'lib/datetime';
-import { ethers } from 'ethers';
-import { isValid, format } from 'date-fns';
+import { formatLocation, formatValue } from '../helpers/spellDiffParsers';
+
 import { SpellDiff as SpellDiffType } from '../../../pages/executive/[proposal-id]';
 
 const CircleIcon = ({ name }) => (
@@ -27,41 +27,9 @@ const CircleIcon = ({ name }) => (
   </Flex>
 );
 
-const parseDiffs = diffs => {
-  //todo
-  return diffs;
-};
-
-const validateDiff = diff => {
-  const contract = diff.decoded_contract ?? (diff.contract || 'N/A');
-  const location = diff.decoded_location ?? (diff.location || 'N/A');
-  const fromVal = diff.decoded_from_val ?? (diff.from_val || 'N/A');
-  const toVal = diff.decoded_to_val ?? (diff.to_val || 'N/A');
-
-  return { contract, location, fromVal, toVal };
-};
-
-const formatLocation = location => {
-  if (location.indexOf(']') !== -1) {
-    const mapKey = location.slice(location.indexOf('[') + 1, location.indexOf(']'));
-    if (ethers.utils.isAddress(mapKey)) console.log('mapkey is address', mapKey);
-    else if (isNaN(parseInt(mapKey))) console.log('mapkey is a contract', mapKey);
-    else console.log('mapkey is an ID', mapKey);
-  }
-  return location;
-};
-
-const formatValue = value => {
-  if (value.indexOf('16') === 0 && isValid(new Date(parseInt(value) * 1000))) {
-    console.log('date', new Date(parseInt(value) * 1000));
-    return { interpreted: format(new Date(parseInt(value) * 1000), 'MM/dd/yyyy') };
-  }
-  return value;
-};
-
 const SpellDiff = ({ diffs }) => {
   return diffs.map((diff, i) => {
-    const { contract, location, fromVal, toVal } = validateDiff(diff);
+    const { contract, location, fromVal, toVal } = diff;
     return (
       <Flex
         key={JSON.stringify(diff[i]) + i}
